core(rename): titlecase file names to match class
