+++ packages/common/core/lib/main.ts
@@ -1,5 +1,4 @@
-export * from './component';
-export * from './root';
-export * from './state';
-export * from './types';
+export * from './ComponentModel';
+export * from './RootModel';
+export * from './StateModel';
 export * from './utils/map';
