--- packages/common/core/lib/types.ts
@@ -1,1 +1,0 @@
-export type $Predicate<TValue> = (value: TValue) => boolean;
